pyproject.toml
[project]
name = "python-tkgm"
version = "1.0.0"
description = "TKGM WFS Data Scraper - Türkiye Tapu ve Kadastro Genel Müdürlüğü parsel verilerini otomatik olarak toplar"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "barisariburnu", email = "barisariburnu@users.noreply.github.com" }
]
dependencies = [
    "requests>=2.31.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.0",
    "shapely>=2.0.2",
    "loguru>=0.7.2",
    "pyproj>=3.7.1",
    "lxml>=6.0.2",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.0.0",
    "types-requests>=2.31.0",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
